+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add display meta property support for tree content models
 - Add active property to field-nav items
 - Add additional security check when uploading files
+- Fix assets image cache check
 
 
 ## 2.4.0 (2023-03-08)